nhamer nhamer casca
The Colette Wave Cooker is an unlockable secondary weapon for the Driller. It is a hand-held microwave emitter composed of various metal parts and a series of electrical inductors. An arm brace secures the grip in the back with the trigger wired directly to the microwave generator. A single amplifying lens is positioned at the front, directing deadly microwave energy towards anything the weapon is aimed at.

The Wave Cooker, similar to Gunner's "Lead Storm" Powered Minigun and Driller's, feeds from a single ammo pool and has an overheating mechanic. This means that heat will be produced for as long as the trigger is held down, which is measured on the weapon's digital display. After firing for an extended period of time and completely filling the heat meter, the Wave Cooker will jam, rendering the weapon inoperable for five seconds.

When using the Wave Cooker, a cylindrical beam area-of-effect is generated and inflicts continuous damage onto any unobstructed target that is within its path from the player's direction. The invisible beam fires in a helical pattern in 24 distinct angles, with VFX only appearing from its point of impact and also has an effective infinite range. The weapon's crosshair attempts to depict the width of the beam based on the distance between the weapon and its impact point, however after a certain amount of distance, the width of the beam appears much smaller than the crosshair due to the perspective of a cylindrical projection with a static diameter. Due to the shape of the projectile, the affected area extremely close to the player and extremely far away does not line up with the crosshair on-screen.

Enemies damaged and killed by the Wave Cooker produces a bubbling visual effect.